continue WIP on headers detection. 7 errror remain

+++ c/csv/reader_fread.h
@@ -130,7 +130,7 @@ private:
   void userOverride();
 
   void detect_column_types();
-  size_t parse_single_line(FreadTokenizer&, bool* bumped);
+  int64_t parse_single_line(FreadTokenizer&, bool* bumped);
 
   friend FreadLocalParseContext;
   friend FreadChunkedReader;
